+++ src/api/router.go
@@ -62,7 +62,7 @@ func newRouterV1(apis *APIs) *gear.Router {
 	routerV1 := gear.NewRouter(gear.RouterOptions{
 		Root: "/v1",
 	})
-	routerV1.Use(tracing.New())
+	routerV1.Use(tracing.New("urbs-console"))
 
 	routerV1.Get("/canary", apis.Canary.Get)
 
@@ -78,12 +78,14 @@ func newRouterAPIV1(apis *APIs) *gear.Router {
 	routerV1 := gear.NewRouter(gear.RouterOptions{
 		Root: "/api/v1",
 	})
-	routerV1.Use(tracing.New())
+	routerV1.Use(tracing.New("urbs-console"))
 	routerV1.Use(middleware.Verify(services))
 
 	// ***** client ******
 	// 读取指定用户的功能配置项，支持条件筛选，数据用于客户端
 	routerV1.Get("/users/settings:unionAll", apis.User.ListSettingsUnionAllClient)
+	// 读取指定用户的标签，支持条件筛选，数据用于客户端
+	routerV1.Get("/users/labels", apis.User.ListLablesForClient)
 
 	checkSuperAdmin := middleware.CheckSuperAdmin(blls)